refactor(etf-detail): merge Price/Returns/CAGR into one tabbed chart
Per follow-up: drop the separate Returns/CAGR chart section and instead
fold Price (line), Returns (bars), and CAGR (bars) into a single tabbed
container — saves vertical space on the ETF detail page.

- New `EtfChartTabs` renders one `<section>` with Price / Returns / CAGR
  top tabs. When Price is selected it embeds `PriceChart` and exposes its
  range buttons (1M / 6M / 1Y / 3Y / 5Y); when Returns or CAGR is selected
  it renders the bar chart with the focal ETF and category-average
  datasets.
- `PriceChart` gains an `embedded` prop. In embedded mode it skips the
  outer `<section>` wrapper and its own title so it can be hosted inside
  EtfChartTabs without nested cards; the stock detail page still uses the
  standalone mode unchanged.
- Drop 3M and YTD from the Returns period set — the original spec only
  asked for 1M / 6M / 1Y / 3Y / 5Y + 10Y / 15Y / 20Y. The API select clause
  drops the corresponding peer-average lookups too.
- Tabs whose data is empty (no priceHistory, or all return/CAGR values
  null) are removed from the tab strip so users can't click into an empty
  panel.
