+++ insights-ui/src/app/api/[spaceId]/etfs-v1/exchange/[exchange]/[etf]/full-render/route.ts
@@ -225,9 +225,7 @@ async function fetchCategoryPeerAnalyzerInfo(spaceId: string, category: string |
       cagr15y: true,
       cagr20y: true,
       return1m: true,
-      return3m: true,
       return6m: true,
-      returnYtd: true,
       return1y: true,
       return3y: true,
       return5y: true,
